move meta settings to a module
